new: add a plugin system (#5)
Plugins allows to inject some code from a so file. This code can be run in multiple
places by a3s, ensuring maximum performance. For now only Token() is defined, which
allowed to hook the issue process, right before the token is returned, acting as a
source modifier, but as code.