new: token revocation
This patch adds a new /revocation API that allows to mark an a3s token as revoked in a namespace.
When the token is revoked, authentication is still OK, but not authorization if checked in the namespace
of the revocation, or one of its children. The revocation objects identifies the token using its ID (jti)
and allows to set an expiration date after which the revocation will be removed.
Revocation are always propagating down.