new: implements importing in subnamespaces
This patchs adds support for importing objects in subnamespaces. For instance

objects:
 - name: a
 - name: b
   namespace: ./subns

If the base namespace is /ns, this would import object "a" in /ns  of the import
and "b" in in /ns/subns.

Subns imports must start with './'

This function is not responsible for ordering of imports, and subnamespaces exist
somehow (either already existing, of created earlier in the import)