new: support for groups
Groups are a new API that allows to define a group of claims and bind them to a name.
The group contains a classic subject.
When the permissions are resolved, the groups with a subject matching the claims
of the user will be retrieved, and then will be injected in the original claims of
the user as "@group:name=[group-name]". This claim can then be used in the authorization
and will apply to all the bearers matching the group's subject.

example:

authorization:
 - subject:
   - - @group:name=managers

group:
 - name: managers
   subject:
   - - email=john@domain.com
   - - email=bob@domain.com

If the token contains one of the two emails as claim, the group will match
and the api policy described above will apply