new/issue: add a way to waive the max validity of a token
This patch adds a way to pass a secret to a3s to waive the max validity
check. Starting a3s with --jwt-waive-validity-secret will allow the
callers of issue to set issue.InputA3S.WaiveValiditySecret. If it
matches the flag, then max validity check will be waived. This is only
for use by a trusted service from a trusted platform.