issue: add --jwt-forbidden-opaque-keys
This flags allows to define a list of opaque keys that are forbidden
from the requets, unless the validity secret is waived.

As opaque data are requested by the user and some modifiers needs to use
trusted opaque data from internal calls, this allows to make sure these
special keys are not coming from the outside world.