fix: use of wrong parameter for autotls client ca
AutoTLS client CA flag is checked to see if not empty, and if so, it is
used to add a client CA to the client pool. However, a typo was making
it read tthe AutoTLSCA instead of AutoTLSClientCA.

This has been unnoticed, because we usually use the same CA for internal
communications.