breaking: authenticator.CheckAuthentication returns idt
Previously, only the identity token claims were returned. It turns out
that there are a lot of cases where the caller was also interested in
other parts of the token, like its ID, expiration etc. The caller would
then need to decode the token again, which is not optimized as
CheckAuthentication just did it internally.

This patch now makes CheckAuthentication return the token.IdentityToken