saml: support url for IDP metadata
This patch adds the property IDPMetadataURL which allows to give a URL
pointing to a IDP Metadata XML file instead of prodividing it directly
in IDPMetadata.

When the IDPMetadataURL is set, the metadata file will be retrieved and
cached when needed to authn a user. The data is cached for 1h.

This allows us to support Microsoft Identity Platform.