new: make included/excluded keys generic
All sources can now set included/excluded keys (not only ldap and saml).
The behavior now also aligns with cloaking. A key matches if is if
prefixed with one the items in the list (previously it was an exact
match on the key).

The name includedClaims and excludedClaims would fit better, however we
want to stay backward compatible.