new: add a way to force checking from MTLS only from header
When a3s is behind a gateway, it is not desirable to use the gateway's
client certificate. There was already a way to pass the verified client
certificate from the Headers, but there was no way to force it.

This patch adds --mtls-header-force-header-key and
--mtls-header-force-heade-value. This means that is the former is equal
to "A" and the later "B", a3s will only check MTLS from the
X-Certificate header if there is another header "A: B" and will skip
looking at the TLSState altogether