new: revocation by subject
Revocation were only used to revoke a single token by its JID. This
patch makes them more powerful by allowing to set a subject instead.
During a revocation check, if the user claims match any active
revocation subject, then the token will be considered as revoked.

Just setting the tokenID still work. If both tokenID and subject are set
in the same revocation, any matching condition will consider the token
revoked.

Also, previously it was only 24h because it was assuming only short lived
tokens. This is not the case anymore. And we need a time for the cleanup
job anyways. 10y should be enough