authorizer: fixed potential cache mixmatch
A recent patch converted a check into a pointer check, needing an
additional nil check. Issue was before when the value was empty we were
generating a key based on a different algorithm. In the conversion if
the nil check did not pass, then the cache key would stay empty and
multiple different cached item would be mixed up